+++ sakura_core/view/CEditView_Diff.cpp
@@ -201,7 +201,7 @@ bool COutputAdapterDiff::OutputA(const ACHAR* pBuf, int size)
 
 	//読み出した文字列をチェックする
 	int j;
-	for( j = 0; j < (int)size/*-1*/; j++ )
+	for( j = 0; j < size/*-1*/; j++ )
 	{
 		if( bLineHead )
 		{
